+++ tests/urlparser.php
@@ -1,0 +1,18 @@
+<?php
+
+require_once(__DIR__.'/../src/parser/url.php');
+
+// TEST
+// - 00
+require_once(__DIR__.'/urlparser_00.php');
+function TestBasics() {
+
+    $test = TestUrlParser_Structure();
+    var_dump($test);
+
+    $test = TestUrlParser_Parsing();
+    var_dump($test);
+
+    $test = TestUrlParser_GetPars();
+    var_dump($test);
+}