+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change log
 ## Current version
-### 1.0.0 - First release
-- new Tool.Structure - node hierarchy class (node mesh); basics and tests
+### 1.0.1 - More tools
+- new Tool.Parser - URL parser; basics and tests
 
 ## History
-- none
+### 1.0.0 - First release
+- new Tool.Structure - node hierarchy class (node mesh); basics and tests